Update README.md with info on SC-E6010
Added info on the SC-E6010. Note: If you read this and you would like to hack it, I have the display. It is still working. However the display itself has some problems and cannot be used as the segments are disturbed probably due to some oxidation. So drop me a note if you want it. First come basis. And of course if you succeed put up a PR so it can be added.